+++ src/main/java/roomescape/dto/CreateReservationRequest.java
@@ -2,5 +2,5 @@ package roomescape.dto;
 
 import java.time.LocalDate;
 
-public record CreateReservationRequest(String name, LocalDate date, int timeId) {
+public record CreateReservationRequest(String name, LocalDate date, long timeId) {
 }
